Fixed Fading Between Maps
Sometimes when going between a shop and the hallway at a fast interval, you can see garbled tiles as the tileset switches. Other times the palette may not fade correctly due to it not being reset.

This adds a palette fade reset, which fixes both problems.

On very rare occasions, the player may still see garbled tiles for a frame, but it's only if they're persistently running between maps. Likely a processing issue, but it's so unnoticeable that this should be good enough.