New Giant Palette Fix
Fixed the weird bug where the giant's palette would resort to the normal, unblended one after battle if the blend/tint was active/wouldn't reapply the effect. Came up with a new solution to hopefully fix that where the giant's palette is just loaded once on callback to the overworld.